+++ app/services/data_export.rb
@@ -97,11 +97,10 @@ module DataExport
       [ customer.id, customer.name, customer.email, customer.phone_number, customer.address,
         customer.city, customer.zip_code, customer.notes, customer.custom_field_1,
         customer.custom_field_2, customer.custom_field_3, customer.custom_field_4,
-        customer.custom_field_5, customer.language, customer.timezone,
-        customer.other_emails, customer.other_phones ]
+        customer.custom_field_5, customer.language, customer.timezone ]
     end
     [ %w[id name email phone_number address city zip_code notes custom_field_1 custom_field_2
-         custom_field_3 custom_field_4 custom_field_5 language timezone other_emails other_phones] ] + rows
+         custom_field_3 custom_field_4 custom_field_5 language timezone] ] + rows
   end
 
   def appointment_statuses_sheet