+++ db/migrate/20260920170000_add_other_contacts_to_users.rb
@@ -1,6 +1,0 @@
-class AddOtherContactsToUsers < ActiveRecord::Migration[8.1]
-  def change
-    add_column :users, :other_emails, :text
-    add_column :users, :other_phones, :text
-  end
-end