+++ test/integration/inbound_messages_test.rb
@@ -32,6 +32,16 @@ class InboundMessagesTest < ActionDispatch::IntegrationTest
     assert_equal "twilio", message.channel
   end
 
+  test "an SMS in E.164 matches a customer whose number was typed in UK local format" do
+    users(:jx).update!(phone_number: "07700 900123")
+    params = { "From" => "+447700900123", "To" => "+15005550006", "Body" => "Running late" }
+    url = "http://www.example.com/messages/inbound/twilio/secrettoken123"
+    post "/messages/inbound/twilio/secrettoken123", params: params,
+         headers: { "X-Twilio-Signature" => twilio_signature(url, params) }
+    assert_response :success
+    assert_equal users(:jx).id, Message.incoming.sole.customer_id
+  end
+
   test "an SMS from a customer fires the customer message notification" do
     Notification.create!(title: "Customer Message Received", event: "customer_message", audiences: %w[provider],
                          channels: %w[email], short_text: "New message from {{Customer Name}}")
@@ -43,16 +53,6 @@ class InboundMessagesTest < ActionDispatch::IntegrationTest
          headers: { "X-Twilio-Signature" => twilio_signature(url, params) }
     assert_response :success
     assert_equal [ users(:zane).email ], Message.outgoing.pluck(:to_address)
-  end
-
-  test "an SMS from a customer's other phone number is theirs" do
-    users(:jx).update!(other_phones: "+447700900450")
-    params = { "From" => "+447700900450", "To" => "+15005550006", "Body" => "Other phone" }
-    url = "http://www.example.com/messages/inbound/twilio/secrettoken123"
-    post "/messages/inbound/twilio/secrettoken123", params: params,
-         headers: { "X-Twilio-Signature" => twilio_signature(url, params) }
-    assert_response :success
-    assert_equal users(:jx).id, Message.incoming.sole.customer_id
   end
 
   test "twilio webhook rejects a bad signature" do